+++ Code/src/main/java/com/yuditsky/advertsik/repository/AdRepository.java
@@ -1,12 +1,10 @@
 package com.yuditsky.advertsik.repository;
 
-import com.yuditsky.advertsik.bean.Ad;
+import com.yuditsky.advertsik.domain.Ad;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AdRepository extends CrudRepository<Ad, Integer> {
-    //List<Ad> findAdByTitle(String title);
+public interface AdRepository extends CrudRepository<Ad, Long> {
     List<Ad> findByTitle(String title);
-
 }
